requirements.txt
Django
channels
channels_redis
django-rest-framework
kiteconnect
django-cors-headers
psycopg2-binary>=2.8
Twisted[tls,http2]
pymongo
dnspython
redis
cryptography
jwt
pytz
docker